Backend/requirements.txt
# CRUD AI-Chat App - Backend Dependencies
# Core FastAPI framework and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database ORM and PostgreSQL driver
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Data validation and serialization
pydantic==2.5.0

# Environment variable management
python-dotenv==1.0.0

# HTTP client for Ollama API
httpx==0.25.2

# CORS middleware (included with FastAPI)
# Future AI integration
# openai==1.3.7  # Uncomment when implementing AI features

# Development and testing (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
